pyproject.toml
[project]
name = "latest_ai_development"
version = "0.1.0"
description = "latest-ai-development using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=0.5.23",
    "crewai-tools>=0.59.0",
    "crewai[tools]>=0.152.0,<1.0.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
latest_ai_development = "latest_ai_development.main:run"
run_crew = "latest_ai_development.main:run"
train = "latest_ai_development.main:train"
replay = "latest_ai_development.main:replay"
test = "latest_ai_development.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
